move jinja filters to common module
